+++ apm-agent-api/src/main/java/co/elastic/apm/api/NoopTransaction.java
@@ -70,6 +70,24 @@ enum NoopTransaction implements Transaction {
         return this;
     }
 
+    @Nonnull
+    @Override
+    public Transaction addCustomContext(String key, String value) {
+        return this;
+    }
+
+    @Nonnull
+    @Override
+    public Transaction addCustomContext(String key, Number value) {
+        return this;
+    }
+
+    @Nonnull
+    @Override
+    public Transaction addCustomContext(String key, boolean value) {
+        return this;
+    }
+
     @Override
     public Transaction setUser(String id, String email, String username) {
         // noop