+++ apm-agent-plugins/apm-servlet-plugin/src/main/java/co/elastic/apm/agent/servlet/ServletTransactionHelper.java
@@ -319,37 +319,39 @@ public class ServletTransactionHelper {
     }
 
     @VisibleForAdvice
-    public static void setTransactionNameByServletClass(String method, @Nullable Class<?> servletClass, StringBuilder transactionName) {
+    public static void setTransactionNameByServletClass(@Nullable String method, @Nullable Class<?> servletClass, StringBuilder transactionName) {
         if (servletClass == null || transactionName.length() > 0) {
             return;
         }
         String servletClassName = servletClass.getName();
         transactionName.append(servletClassName, servletClassName.lastIndexOf('.') + 1, servletClassName.length());
-        transactionName.append('#');
-        switch (method) {
-            case "DELETE":
-                transactionName.append("doDelete");
-                break;
-            case "HEAD":
-                transactionName.append("doHead");
-                break;
-            case "GET":
-                transactionName.append("doGet");
-                break;
-            case "OPTIONS":
-                transactionName.append("doOptions");
-                break;
-            case "POST":
-                transactionName.append("doPost");
-                break;
-            case "PUT":
-                transactionName.append("doPut");
-                break;
-            case "TRACE":
-                transactionName.append("doTrace");
-                break;
-            default:
-                transactionName.append(method);
+        if (method != null) {
+            transactionName.append('#');
+            switch (method) {
+                case "DELETE":
+                    transactionName.append("doDelete");
+                    break;
+                case "HEAD":
+                    transactionName.append("doHead");
+                    break;
+                case "GET":
+                    transactionName.append("doGet");
+                    break;
+                case "OPTIONS":
+                    transactionName.append("doOptions");
+                    break;
+                case "POST":
+                    transactionName.append("doPost");
+                    break;
+                case "PUT":
+                    transactionName.append("doPut");
+                    break;
+                case "TRACE":
+                    transactionName.append("doTrace");
+                    break;
+                default:
+                    transactionName.append(method);
+            }
         }
     }